+++ test/unit/lwrps/base_config_spec.rb
@@ -3,6 +3,84 @@ require 'json'
 
 describe "sensu_base_config" do
 
+  let(:single_broker_config) do
+    {
+      "host" => "10.0.0.6",
+      "port" => 5671,
+      "vhost" => "/sensu",
+      "user" => "sensu",
+      "password" => "secret",
+      "heartbeat" => 30,
+      "prefetch" => 50,
+      "ssl" => {
+        "cert_chain_file" => "/etc/sensu/ssl/cert.pem",
+        "private_key_file" => "/etc/sensu/ssl/key.pem"
+      },
+      "single_broker" => true
+    }
+  end
+
+  let(:multiple_broker_config) do
+    {
+      "hosts" => [ "10.0.0.6", "10.0.0.7", "10.0.0.8" ],
+      "host" => "192.168.1.1",
+      "port" => 5671,
+      "vhost" => "/sensu",
+      "user" => "sensu",
+      "password" => "secret",
+      "heartbeat" => 30,
+      "prefetch" => 50,
+      "ssl" => {
+        "cert_chain_file" => "/etc/sensu/ssl/cert.pem",
+        "private_key_file" => "/etc/sensu/ssl/key.pem"
+      }
+    }
+  end
+
+  let(:multiple_broker_json) do
+    [
+      {
+        "host" => "10.0.0.6",
+        "port" => 5671,
+        "vhost" => "/sensu",
+        "user" => "sensu",
+        "password" => "secret",
+        "heartbeat" => 30,
+        "prefetch" => 50,
+        "ssl" => {
+          "cert_chain_file" => "/etc/sensu/ssl/cert.pem",
+          "private_key_file" => "/etc/sensu/ssl/key.pem"
+        }
+      },
+      {
+        "host" => "10.0.0.7",
+        "port" => 5671,
+        "vhost" => "/sensu",
+        "user" => "sensu",
+        "password" => "secret",
+        "heartbeat" => 30,
+        "prefetch" => 50,
+        "ssl" => {
+          "cert_chain_file" => "/etc/sensu/ssl/cert.pem",
+          "private_key_file" => "/etc/sensu/ssl/key.pem"
+        }
+      },
+      {
+        "host" => "10.0.0.8",
+        "port" => 5671,
+        "vhost" => "/sensu",
+        "user" => "sensu",
+        "password" => "secret",
+        "heartbeat" => 30,
+        "prefetch" => 50,
+        "ssl" => {
+          "cert_chain_file" => "/etc/sensu/ssl/cert.pem",
+          "private_key_file" => "/etc/sensu/ssl/key.pem"
+        }
+      }
+    ]
+  end
+
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       :platform => "ubuntu",
@@ -10,9 +88,9 @@ describe "sensu_base_config" do
       :step_into => ['sensu_base_config', 'sensu_json_file']
     ) do |node|
       node.set["sensu"]["transport"]["chefspec"] = true
-      node.set["sensu"]["rabbitmq"]["chefspec"] = true
       node.set["sensu"]["redis"]["chefspec"] = true
       node.set["sensu"]["api"]["chefspec"] = true
+      node.set["sensu"]["rabbitmq"] = single_broker_config
     end.converge("sensu::default")
   end
 
@@ -23,11 +101,34 @@ describe "sensu_base_config" do
   end
 
   context "base configuration is derived from node attributes" do
-    %w( transport rabbitmq redis api ).each do |kw|
+    %w[ transport redis api ].each do |kw|
       it "#{kw} node attributes are present in base configuration" do
         content = JSON.parse(base_config_json.content)
         expect(content[kw]["chefspec"]).to eq(true)
       end
+    end
+  end
+
+  context "single rabbitmq host provided" do
+    it "yields a rabbitmq array with a single hash" do
+      content = JSON.parse(base_config_json.content)
+      expect(content["rabbitmq"].is_a?(Array)).to eq(true)
+      expect(content["rabbitmq"][0].is_a?(Hash)).to eq(true)
+      expect(content["rabbitmq"][0]).to eq(single_broker_config)
+    end
+  end
+
+  context "multiple rabbitmq hosts provided" do
+
+    before do
+      chef_run.node.set["sensu"]["rabbitmq"] = multiple_broker_config
+      chef_run.converge("sensu::default")
+    end
+
+    it "yields a rabbitmq array containing multiple brokers" do
+      content = JSON.parse(base_config_json.content)
+      expect(content["rabbitmq"].is_a?(Array)).to eq(true)
+      expect(content["rabbitmq"]).to eq(multiple_broker_json)
     end
   end