+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to track changes made in each version of the sensu cookbook.
 
+## Unreleased
+
+### Features
+
+* The `sensu_base_config` provider now honors `node["sensu"]["rabbitmq"]["hosts"]` attribute, providing an array of hosts to use for configuring rabbitmq transport with multiple brokers. When empty, falls back to existing `node["sensu"]["rabbitmq"]["host"]` attribute.
+
 ## 2.12.0 - 2016-03-14
 
 ### Project changes